+++ inst/essais/clippedThreeCylinders.R
@@ -22,9 +22,9 @@ cyl3 <- as.tmesh3d(cylinder3d(path3, radius = 1, sides = 360, closed = -2))
 mesh1 <- cgalMesh$new(cyl1)
 mesh2 <- cgalMesh$new(cyl2)
 mesh3 <- cgalMesh$new(cyl3)
-mesh1$assignFaceColors("firebrick")
-mesh2$assignFaceColors("forestgreen")
-mesh3$assignFaceColors("midnightblue")
+mesh1$assignFaceColors("hotpink")
+mesh2$assignFaceColors("chartreuse")
+mesh3$assignFaceColors("gold")
 
 
 .meshes1 <- mesh1$clip(mesh2, TRUE)
@@ -34,12 +34,12 @@ rmesh1 <- mesh1$getMesh()
 shade3d(rmesh1, meshColor = "faces")
 
 
-blue <- which(mesh1$getFaceColors() == "midnightblue")
+blue <- which(mesh1$getFaceColors() == "hotpink")
 mm1 <- mesh1$filterMesh(blue)
 mblue <- mm1[[1]]
 m2    <- mm1[[2]]
 
-red <- which(m2$getFaceColors() == "firebrick")
+red <- which(m2$getFaceColors() == "chartreuse")
 mm2 <- m2$filterMesh(red)
 mred   <- mm2[[1]]
 mgreen <- mm2[[2]]
@@ -55,9 +55,9 @@ rm3 <- mblue$getMesh()
 
 open3d(windowRect = 50 + c(0, 0, 512, 512))
 view3d(30, 30, zoom = 0.7)
-shade3d(rm1, meshColor = "faces", specular = "darkred")
-shade3d(rm2, meshColor = "faces", specular = "darkgreen")
-shade3d(rm3, meshColor = "faces", specular = "darkblue")
+shade3d(rm1, meshColor = "faces", specular = "pink")
+shade3d(rm2, meshColor = "faces", specular = "limegreen")
+shade3d(rm3, meshColor = "faces", specular = "yellow")
 
 
 # # -- if you want an animation