+++ src/MODULE.cpp
@@ -72,6 +72,7 @@ RCPP_MODULE(class_CGALmesh) {
     .method("print", &CGALmesh::print)
     .method("removeSelfIntersections", &CGALmesh::removeSelfIntersections)
     .method("reverseFaceOrientations", &CGALmesh::reverseFaceOrientations)
+    .method("sampleMesh", &CGALmesh::sampleMesh)
     .method("sharpEdges", &CGALmesh::sharpEdges)
     .method("Sqrt3Subdivision", &CGALmesh::Sqrt3Subdivision)
     .method("subtract", &CGALmesh::subtract)